[node] Modularize the core tree module
(needs some debugging I believe?)
